requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
cycler==0.11.0
discord.py==1.7.3
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.2
multidict==6.0.2
numpy==1.22.3
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.2.0
yarl==1.9.2
